Create ContentView.swift
In this game of rock, paper, scissors, the application will randomly choose one of three possible moves. After choosing a move, the app will then challenge the user to either win or lose the current engagement.

It will be the players objective to quickly and successfully decide which of their three options, rock, paper, or scissors, will fulfill the parameters of the game.